+++ Commands/eventsCommands.py
@@ -29,8 +29,12 @@ def birthdayList(bot, update, args=None):
 		formatedEventList = ''
 		lap = 1
 		for event in eventList:
-			date = timezone('UTC').localize(datetime.strptime(event['start']['date'], '%Y-%m-%d')).astimezone(timezone('Europe/Madrid')).strftime("%d-%m-")
-			formatedEventList += (str(lap) + '. ' + event['summary'].split('|')[0] + ': ' + date + event['summary'].split('|')[2] + '\n')
+			date = timezone('UTC').localize(datetime.strptime(event['start']['date'], '%Y-%m-%d')).astimezone(timezone('Europe/Madrid')).strftime("%d-%m")
+			if bd.user_id != bd.chatIDDeveloper and bd.user_id not in bd.adminTelegramId:
+				formatedEventList += (str(lap) + '. ' + date + ': ' + event['summary'].split('|')[0] + '\n')
+			else:
+				formatedEventList += (str(lap) + '. ' + date + '-' + event['summary'].split('|')[2] + ': ' + event['summary'].split('|')[0] + '\n')
+
 			lap += 1
 		if bd.user_id == bd.chatIDDeveloper or bd.user_id in bd.adminTelegramId:
 			bot.sendMessage(chat_id=bd.chat_id, text=formatedEventList, reply_to_message_id=bd.message.message_id)